refactor: rename parser module to wiki-parser
